% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_MLE_policy.R
\name{act_MLE_policy}
\alias{act_MLE_policy}
\title{The act_MLE_policy Function}
\usage{
act_MLE_policy(
  x_t,
  mu_t,
  sig_t,
  rf = 0.02,
  x0 = 1,
  z = 1.4,
  invest_hrzn = 1,
  exploratory = F,
  lambda = NULL,
  current_t = NULL
)
}
\arguments{
\item{x_t}{(Required, numeric) current wealth}

\item{mu_t}{(Required, numeric) current MLE estimate of mu}

\item{sig_t}{(Required, numeric) current MLE estimate of sigma}

\item{rf}{(Optional, numeric) riskfree interest rate, default to 0.02}

\item{x0}{(Optional, numeric) initial wealth, detault to 1}

\item{z}{(Optional, numeric) target terminal wealth, detault to 1.4}

\item{invest_hrzn}{(Optional, numeric) investment horizon, total length of the investment
period in years, default to 1.}

\item{exploratory}{(Optional, logical) default to False. If set to True, then apply the
optimal Exploratory Mean-Variance (EMV) policy by sampling \eqn{u_t \sim N(u_t^*, \frac{\lambda}{2\sigma^2}e^{-\rho^2(T-t)})},
with all of \eqn{(\rho, \mu, \sigma)} replaced with their MLE estimates.}

\item{lambda}{(Optional, numeric) the exploratory weight, default to NULL, only use it if
`exploratory == TRUE`}

\item{current_t}{(Optional, numeric) the current time, default to NULL, only use it if
`exploratory == TRUE`.}
}
\value{
The amount of money invest in stock (a numerical value of length 1).
}
\description{
`act_MLE_policy()` make an investment decision following the classical Mean-Variance policy.
Suppose the current wealth at time t is \eqn{x_t}, then the classical optimal policy is to
invest \eqn{u_t^* = -\frac{\rho}{\sigma}(x-\omega)} in the stock and invest the rest
\eqn{x_t - u_t} in the bond, where \eqn{\rho := \frac{\mu - r}{\sigma}} is the Sharpe ratio
and \eqn{r} is the riskfree interest rate.
}
\examples{
# Take a non-exploratory action
act_MLE_policy(x_t = 0, mu_r=0.3, sig_t=0.2)

# Take an exploratory action
act_MLE_policy(x_t = 0, mu_r=0.3, sig_t=0.2, exploratory=T, lambda=2, current_t=1/252)
}
