% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockPath.R
\name{getStockPath}
\alias{getStockPath}
\title{The `getStockPath` Function}
\usage{
getStockPath(
  mu,
  sigma,
  S0 = 1,
  dt = 1/252,
  invest_hrzn = 1,
  mwindow_size = 100
)
}
\arguments{
\item{mu}{(Required, numeric) the drift parameter in the stock price process dynamic.}

\item{sigma}{(Required, numeric) the diffusion parameter in the stock price process dynamic.}

\item{S0}{(Optional, numeric) the initial stock price, default to 1.}

\item{dt}{(Optional, numeric) the discretization of continuous time, default to 1/252,
representing daily discretization. If set to 1/12, then this is monthly discretization.}

\item{invest_hrzn}{(Optional, numeric) investment horizon, total length of the investment
period in years, default to 1.}

\item{mwindow_size}{(Optional, numeric) moving window size, default to 100 and must be non-negative.}
}
\value{
A simulated path of stock prices of length `invest_hrzn/dt + mwindow_size`.
}
\description{
`getStockPath()` simulates a path of stock price process, following a geometric Brownian Motion.
}
\details{
The stock price process \eqn{\{S_t\}_{0\leq t \leq T}} followng a geometric Brownian Motion with dynamic:
\eqn{d S_t = S_t (\mu dt + \sigma dW_t)}, where \eqn{\{W_t\}_{0\leq t \leq T}} is a 1-dim Brownian Motion.
}
\examples{
# Simulate 100 paths of stocks prices.
for (n in 1:100){
  St = simulate_stock_path(mu=mu, sigma=sig, dt=dt, invest_hrzn = 1, mwindow_size = 0)
  if (n == 1){
  plot(St, type = "l", ylim = c(0,2), col="gray",
      ylab="Stock Price ($)", xlab = "Time (days)")
  }
  else{
  lines(St, col="gray")
  }
}
}
