% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerminalWealth.R
\name{getTerminalWealth}
\alias{getTerminalWealth}
\alias{plot_terminalwealth}
\title{The `getTerminalWealth` Function}
\usage{
getTerminalWealth(
  St,
  policy = "MLE classic",
  dt = 1/252,
  rf = 0.02,
  x0 = 1,
  z = 1.4,
  invest_hrzn = 1,
  lambda = 2,
  mwindow_size = 100,
  phi1 = 1,
  phi2 = 0.5,
  w = 1,
  constrained_invest = FALSE,
  max_invest = 5
)
}
\arguments{
\item{St}{a stock price path}

\item{policy}{the investment policy, default to 'MLE classic', can also take one of (MLE exploratory', 'EMV')}

\item{dt}{(Optional, numeric) the discretization of continuous time, default to 1/252,}

\item{rf}{(Optional, numeric) riskfree interest rate, default to 0.02}

\item{x0}{(Optional, numeric) initial wealth, detault to 1}

\item{z}{(Optional, numeric) target terminal wealth, detault to 1.4}

\item{invest_hrzn}{(Optional, numeric) investment horizon, total length of the investment
period in years, default to 1.}

\item{lambda}{(Optional, numeric) the exploratory weight, default to 2}

\item{mwindow_size}{(Optional, numeric) moving window size, default to 100 and must be non-negative.}

\item{phi1}{(Optional, numeric) a parameter in the optimal EMV policy distribution, default to 1}

\item{phi2}{(Optional, numeric) a parameter in the optimal EMV policy distribution, default to 0.5}

\item{w}{(Optional, numeric) Lagrange multiplier, default to 1}

\item{constrained_invest}{(Optional, logical) default to FALSE, if set to TRUE, then investment
is constrained so that abs(investment) <= `max_invest`}

\item{max_invest}{(Optional, numeric) if `constrained_invest` is set to TRUE, then investment
is constrained so that abs(investment) <= max_invest}
}
\value{
A named list containing:
\itemize{
 \item{"wealth": }{the path of wealth over the investment period, a vector of length `invest_hrzn`/`dt`}
 \item{"controls": }{the path of investmen decisions over the investment period, a vector of length `invest_hrzn`/`dt` - 1}
 \item{"terminal_wealth": }{the wealth at the end of the investment period, i.e., the last value of `wealth`}
}
}
\description{
Given a path of stock price, make investment over time following either the Exploratory
Mean-Variance policy or the classical non-exploratory Mean-Variance (MV) policy with
Maximum Likelihood Estimation. The terminal wealth is the last portfolio value at the
end of the investment period.
}
\examples{
# First get a stock path
St = getStockPath(mu=-0.3, sigma=0.2, dt=1/252, invest_hrzn = 1, mwindow_size = 0)

# Then, trade following the MLE classic policy and compute the terminal wealth
library(zoo)
getTerminalWealth(St)

}
