% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_terminalwealth.R
\name{show_terminalwealth}
\alias{show_terminalwealth}
\alias{plot_terminalwealth}
\alias{table_terminalwealth}
\title{Illustration of Terminal Wealth}
\usage{
plot_terminalwealth(terminal_wealth, plot_type, ...)

table_terminalwealth(terminal_wealth, table_type, rf = 0.02, ...)
}
\arguments{
\item{terminal_wealth}{(Required, numeric) the terminal wealth from 1 or multiple polices, can take either a vector or a
matrix with multiple columns.}

\item{plot_type}{(Required, string) plot type, take value from ('boxplot', 'density', 'ecdf', 'qqplot')}

\item{table_type}{the table type can be one of ('summary', 'quantile')}

\item{rf}{riskfree interest rate, default to 0.02}
}
\description{
`plot_terminalwealth` plots the boxplot/density plot/ECDF plot/qqplot of the terminal wealth.

`table_terminalwealth` tabularizes the summary and quantiles of the terminal wealth.
}
